fix:[YN] fix error text input validation

+++ lib/homeAdmin/presentation/pages/aksi_input_sampah_page.dart
@@ -162,7 +162,7 @@ class _AksiInputSampahPageState extends State<AksiInputSampahPage> {
                                             items.add({
                                               "id":
                                                   state.data[x].products[y].id,
-                                              "quantity": int.parse(
+                                              "quantity": double.parse(
                                                   _listControllerText[x][y]
                                                       .text)
                                             });
